added error recovery logic for contract creation, fixed bugs with HttpConnection timeout

+++ app/src/main/java/ch/uzh/ifi/csg/contract/service/contract/FileManager.java
@@ -99,7 +99,7 @@ public class FileManager implements ContractManager, AccountManager {
             new File(path).delete();
         }
 
-        if(toDelete.getUserProfile().getProfileImagePath() != null)
+        if(toDelete.getUserProfile() != null && toDelete.getUserProfile().getProfileImagePath() != null)
             new File(toDelete.getUserProfile().getProfileImagePath()).delete();
 
         accountMap.get(account).getContracts().remove(contractAddress);
